backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
pydantic>=2.6.0
python-dotenv>=1.0.0
geopy>=2.4.0